Refactor timeouts
