+++ wat/all_imports.wat
@@ -16,38 +16,38 @@
     (import "lunatic::message" "push_udp_socket" (func (param i64) (result i64)))
     (import "lunatic::message" "take_udp_socket" (func (param i64) (result i64)))
     (import "lunatic::message" "send" (func (param i64)))
-    (import "lunatic::message" "send_receive_skip_search" (func (param i64 i32) (result i32)))
-    (import "lunatic::message" "receive" (func (param i32 i32 i32) (result i32)))
+    (import "lunatic::message" "send_receive_skip_search" (func (param i64 i64) (result i32)))
+    (import "lunatic::message" "receive" (func (param i32 i32 i64) (result i32)))
 
     (import "lunatic::timer" "send_after" (func (param i64 i64) (result i64)))
     (import "lunatic::timer" "cancel_timer" (func (param i64) (result i32)))
 
-    (import "lunatic::networking" "resolve" (func (param i32 i32 i32 i32) (result i32)))
+    (import "lunatic::networking" "resolve" (func (param i32 i32 i64 i32) (result i32)))
     (import "lunatic::networking" "drop_dns_iterator" (func (param i64)))
     (import "lunatic::networking" "resolve_next" (func (param i64 i32 i32 i32 i32 i32) (result i32)))
     (import "lunatic::networking" "tcp_bind" (func (param i32 i32 i32 i32 i32 i32) (result i32)))
     (import "lunatic::networking" "drop_tcp_listener" (func (param i64)))
     (import "lunatic::networking" "tcp_local_addr" (func (param i64 i32) (result i32)))
     (import "lunatic::networking" "tcp_accept" (func (param i64 i32 i32) (result i32)))
-    (import "lunatic::networking" "tcp_connect" (func (param i32 i32 i32 i32 i32 i32 i32) (result i32)))
+    (import "lunatic::networking" "tcp_connect" (func (param i32 i32 i32 i32 i32 i64 i32) (result i32)))
     (import "lunatic::networking" "drop_tcp_stream" (func (param i64)))
     (import "lunatic::networking" "clone_tcp_stream" (func (param i64) (result i64)))
-    (import "lunatic::networking" "tcp_write_vectored" (func (param i64 i32 i32 i32 i32) (result i32)))
-    (import "lunatic::networking" "tcp_read" (func (param i64 i32 i32 i32 i32) (result i32)))
+    (import "lunatic::networking" "tcp_write_vectored" (func (param i64 i32 i32 i32) (result i32)))
+    (import "lunatic::networking" "tcp_read" (func (param i64 i32 i32 i32) (result i32)))
     (import "lunatic::networking" "tcp_flush" (func (param i64 i32) (result i32)))
     (import "lunatic::networking" "udp_bind" (func (param i32 i32 i32 i32 i32 i32) (result i32)))
     (import "lunatic::networking" "drop_udp_socket" (func (param i64)))
     (import "lunatic::networking" "udp_local_addr" (func (param i64 i32) (result i32)))
-    (import "lunatic::networking" "udp_receive" (func (param i64 i32 i32 i32 i32) (result i32)))
-    (import "lunatic::networking" "udp_receive_from" (func (param i64 i32 i32 i32 i32 i32) (result i32)))
-    (import "lunatic::networking" "udp_connect" (func (param i64 i32 i32 i32 i32 i32 i32 i32) (result i32)))
+    (import "lunatic::networking" "udp_receive" (func (param i64 i32 i32 i32) (result i32)))
+    (import "lunatic::networking" "udp_receive_from" (func (param i64 i32 i32 i32 i32) (result i32)))
+    (import "lunatic::networking" "udp_connect" (func (param i64 i32 i32 i32 i32 i32 i64 i32) (result i32)))
     (import "lunatic::networking" "clone_udp_socket" (func (param i64) (result i64)))
     (import "lunatic::networking" "set_udp_socket_broadcast" (func (param i64 i32)))
     (import "lunatic::networking" "get_udp_socket_broadcast" (func (param i64) (result i32)))
     (import "lunatic::networking" "set_udp_socket_ttl" (func (param i64 i32)))
     (import "lunatic::networking" "get_udp_socket_ttl" (func (param i64) (result i32)))
-    (import "lunatic::networking" "udp_send_to" (func (param i64 i32 i32 i32 i32 i32 i32 i32 i32 i32) (result i32)))
-    (import "lunatic::networking" "udp_send" (func (param i64 i32 i32 i32 i32) (result i32)))
+    (import "lunatic::networking" "udp_send_to" (func (param i64 i32 i32 i32 i32 i32 i32 i32 i32) (result i32)))
+    (import "lunatic::networking" "udp_send" (func (param i64 i32 i32 i32) (result i32)))
 
     (import "lunatic::process" "compile_module" (func (param i32 i32 i32) (result i32)))
     (import "lunatic::process" "drop_module" (func (param i64)))
@@ -89,7 +89,7 @@
     (import "lunatic::distributed" "module_id" (func (result i64)))
     (import "lunatic::distributed" "spawn" (func (param i64 i64 i64 i32 i32 i32 i32 i32) (result i32)))
     (import "lunatic::distributed" "send" (func (param i64 i64)))
-    (import "lunatic::distributed" "send_receive_skip_search" (func (param i64 i64 i32) (result i32)))
+    (import "lunatic::distributed" "send_receive_skip_search" (func (param i64 i64 i64) (result i32)))
 
     (func (export "hello") nop)
 )
